LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-form-traverse-expected.txt



FAIL reloading a page created from form submission results in formData of null, not the original form data assert_equals: expected (object) null but got (undefined) undefined
